Added: Items adapter, reviews adapter, add item fragment, leave review fragment, complete transaction fragment

+++ app/src/main/java/com/example/nestswap/ItemDetailsFragment.kt
@@ -41,7 +41,7 @@ class ItemDetailsFragment : DialogFragment() {
         binding.btnRent.setOnClickListener {
             dismiss()
 
-            val paymentFragment = PaymentFragment.newInstance(item?.name, item?.price ?: 0)
+            val paymentFragment = PaymentFragment.newInstance(item?.name, (item?.price ?: 0) as Int)
 
             paymentFragment.show(parentFragmentManager, "PaymentDialog")
         }
